requirements.txt
amqp==5.0.6
anyio==3.2.1
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.6.4.0
bitcoin==1.1.39
bleach==3.3.1
blockcypher==1.0.92
boto3==1.17.95
botocore==1.20.95
celery==5.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
device-detector==0.10
dj-database-url==0.5.0
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==1.1.0
django-gm2m==1.1.1
django-guardian==2.4.0
django-heroku==0.3.1
django-model-utils==4.1.1
django-storages==1.11.1
django-user-visit==0.4.1
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
dlib==19.22.0
easy-cache==1.1.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.1.0
idna==2.10
ipykernel==6.0.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
kaggle==1.5.12
kiwisolver==1.3.1
kombu==5.1.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
opencv-python==4.5.2.52
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.19.0
python-mimeparse==1.6.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qrcode==7.2
regex==2021.4.4
requests==2.25.1
requests-unixsocket==0.2.0
s3transfer==0.4.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
whitenoise==5.2.0
